Move BusineeMetadataRecord from protot to data2/metadata/dataset package.
