+++ cdap-data-fabric/src/test/java/co/cask/cdap/data2/metadata/dataset/BusinessMetadataDatasetTest.java
@@ -18,7 +18,6 @@ package co.cask.cdap.data2.metadata.dataset;
 import co.cask.cdap.api.dataset.DatasetProperties;
 import co.cask.cdap.data2.datafabric.dataset.DatasetsUtil;
 import co.cask.cdap.data2.dataset2.DatasetFrameworkTestUtil;
-import co.cask.cdap.proto.BusinessMetadataRecord;
 import co.cask.cdap.proto.Id;
 
 import co.cask.cdap.proto.ProgramType;
